Makes SwiftIterableFlutterPlugin class conform to IterableURLDelegate protocol so that it can receive and propagate Iterable push notifications that are configured with the "Open URL" action.
With this change in place, the openedNotificationHandler in the Flutter application will now also be called for Iterable push notifications that are set-up in Iterable as having the Open action behaviour type set to "Open URL".
Prior to this change, the Flutter application would not be called when configured as "Open URL" so deep-link use cases were not supported.